Modified the program's output. Embedded the SAT solver result.
The program now automatically validates the output generated using the
specified SAT solver's binary. The output includes the SAT solver's
result and out/error logs as well as the mathematical forms of the
input expression and the reverse mapping from the SAT solver's output to
the mathematical expression of the literal that it represents.